Changed controller name and added getOne, getAll and editNurse
