+++ server.js
@@ -4,7 +4,7 @@ import { connect } from './includes/database.js';
 
 import authRouter from './routes/auth.js';
 import adminRouter from './routes/admin.js';
-import tasksRouter from './routes/tasks.js';
+import nurseRouter from './routes/nurse.js';
 
 const app = express();
 const port = process.env.PORT || 5000;
@@ -14,8 +14,8 @@ app.use(express.json());
 
 app.use('/api/auth', authRouter);
 app.use('/api/admin', adminRouter)
-app.use('/api/tasks', tasksRouter)
-// Database
+app.use('/api/nurse', nurseRouter)
+    // Database
 connect();
 
 app.listen(port, () => {
